+++ lib/vagrant/bundler.rb
@@ -77,6 +77,13 @@ module Vagrant
       Gem.clear_paths
     end
 
+    # Removes any temporary files created by init
+    def deinit
+      File.unlink(ENV["BUNDLE_APP_CONFIG"]) rescue nil
+      File.unlink(ENV["BUNDLE_CONFIG"]) rescue nil
+      File.unlink(ENV["GEMFILE"]) rescue nil
+    end
+
     # Installs the list of plugins.
     #
     # @param [Hash] plugins